wip: add typescript
