+++ .gitignore
@@ -3,3 +3,4 @@ node_modules/**/*
 npm-debug.*
 .DS_Store
 package-lock.json
+out
