+++ package.json
@@ -1,11 +1,15 @@
 {
-  "main": "node_modules/expo/AppEntry.js",
+  "main": "src/App.tsx",
   "private": true,
   "scripts": {
     "prettier": "prettier --check \"**/*.js\"",
     "prettier-fix": "prettier --write \"**/*.js\"",
     "lint": "npm run prettier",
-    "test": "npm run lint"
+    "test": "npm run lint",
+    "clean": "rimraf out",
+    "build": "npm run clean && tsc",
+    "buildAndWatch": "npm run build -- -w",
+    "start": "npm run build && concurrently \"npm run buildAndWatch\" \"npm run start\""
   },
   "dependencies": {
     "expo": "^32.0.0",
@@ -16,6 +20,15 @@
     "react-navigation": "^3.0.9"
   },
   "devDependencies": {
-    "prettier": "^1.16.4"
+    "@types/expo": "^32.0.2",
+    "@types/react": "^16.8.3",
+    "@types/react-native": "^0.57.36",
+    "@types/react-navigation": "^3.0.2",
+    "concurrently": "^4.1.0",
+    "expo-cli": "^2.10.1",
+    "prettier": "^1.16.4",
+    "rimraf": "^2.6.3",
+    "tslint": "^5.12.1",
+    "typescript": "^3.3.3"
   }
 }
