+++ src/data/kana/hiragana.ts
