+++ src/data/kana/index.ts
