+++ src/screens/game/Diamonds.tsx
