+++ src/screens/game/Gem.tsx
