+++ src/screens/kanaSelection/index.tsx
