+++ src/screens/pause/index.tsx
